Added Bookmarks file presense checker
- Also added funcionality to extract bookmarks from chrome, chromium and brave without needing jq command
- exploring jq command

+++ main.sh
@@ -44,29 +44,50 @@ if [ $# -eq 0 ]
         
 
         ## Find if BookMarks File Exist
-
-        chrome_bookmark='/Users/gokulsreekumar/Library/Application Support/Google/Chrome/Default/Bookmarks'
-        chromium_bookmark='/Users/gokulsreekumar/Library/Application Support/Chromium/Default/Bookmarks'
-        brave_bookmark="/Users/gokulsreekumar/Library/Application\ Support/BraveSoftware/Brave-Browser/Default/Bookmarks"
+        chrome_bookmark=$(find "/Users/gokulsreekumar/Library/Application Support/Google/Chrome" -name Bookmarks | head -1)
+        chromium_bookmark=$(find "/Users/gokulsreekumar/Library/Application Support/Chromium" -name Bookmarks | head -1)
+        brave_bookmark=$(find "/Users/gokulsreekumar/Library/Application Support/BraveSoftware" -name Bookmarks | head -1)
 
         if [ $chrome_flag -eq 1 ] 
             then
-                cat "$chrome_bookmark" | grep url | grep -v type | awk '{print($2)}' > chrome_bookmarks.txt
+                if [ -f "$chrome_bookmark" ] 
+                    then
+                        cat "$chrome_bookmark" | grep url | grep -v type | awk '{print($2)}' > chrome_bookmarks.txt
+                fi
         fi
-        
-        
+
         if [ $chromium_flag -eq 1 ] 
-            then
-                cat "$chromium_bookmark" | grep url | grep -v type | awk '{print($2)}' > chromium_bookmarks.txt
+            then 
+                if [ -f "$chromium_bookmark" ] 
+                    then
+                        cat "$chromium_bookmark" | grep url | grep -v type | awk '{print($2)}' > chromium_bookmarks.txt
+                fi
         fi
-        
-        find 
-        
+
         if [ $brave_flag -eq 1 ] 
-            then
-                cat "$brave_bookmark" | grep url | grep -v type | awk '{print($2)}' > brave_bookmarks.txt
+            then 
+                if [ -f "$brave_bookmark" ] 
+                    then
+                        cat "$brave_bookmark" | grep url | grep -v type | awk '{print($2)}' > brave_bookmarks.txt
+                fi
         fi
+
+
+
+        # chrome_bookmark='/Users/gokulsreekumar/Library/Application Support/Google/Chrome/Default/Bookmarks'
+        # chromium_bookmark='/Users/gokulsreekumar/Library/Application Support/Chromium/Default/Bookmarks'
+        # brave_bookmark="/Users/gokulsreekumar/Library/Application\ Support/BraveSoftware/Brave-Browser/Default/Bookmarks"
+
+
+        # if [ $chrome_flag -eq 1 ] 
+        #     then
+        #         cat "$chrome_bookmark" | grep url | grep -v type | awk '{print($2)}' > chrome_bookmarks.txt
+        # fi
         
+        # if [ $chromium_flag -eq 1 ] 
+        #     then
+        #         cat "$chromium_bookmark" | grep url | grep -v type | awk '{print($2)}' > chromium_bookmarks.txt
+        # fi
 
-
+    
 fi
